009 07:08 QLThanhVien + Config sidebar

+++ database/migrations/2024_07_11_160536_add_user_catalouge_id_to_users.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('user_catalogue_id')->default(2);
+            $table->bigInteger('user_catalogue_id')->default(2);
         });
     }
